test: disable no-huge where it's not necessary
In tests where no-shconf flag is used, no-huge is also passed due to
compatibility with FreeBSD system, as described in: b5d878e6d.
However on Linux systems with RTE_IOVA_PA (lack of or an incompatible
IOMMU) this causes issues since hugepages are required by EAL.
Therefore replace all occurrences of no_huge which don't actually test
the no-huge logic with a execution environment conditional
no_huge_compat to indicate that it is passed as a compatibility flag,
not as a requirement for a test itself.

Note that checkpatch is complaining about argvX arrays not being static
const. This patch doesn't change that to not add confusion.

Signed-off-by: Stanislaw Kardach <kda@semihalf.com>
Fixes: b5d878e6db56 ("test: fix EAL flags autotest on FreeBSD")
Cc: stable@dpdk.org
Signed-off-by: 0-day Robot <robot@bytheb.org>