examples/l3fwd: fib skip tx queue drain on first iteration
The commit a8f8b672d575 ("examples/l3fwd: skip Tx queue drain on first
iteration") implemented a change to the em and lpm lookup methods to
prevent the TX queue drain running in the first iteration of their
packet processing loops. This patch introduces this change into the
fib lookup method, this was done to keep the fib code aligned with
the em and lpm code.

Signed-off-by: Conor Walsh <conor.walsh@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>