lib: fix socket id type
The variable type of the current socket ID is uint8_t. When traversing all
nodes, the socket ID is compared with RTE_MAX_NUMA_NODES. Since
RTE_MAX_NUMA_NODES has not been verified, it may be larger than UCHAR_MAX
theoretically. This would lead to an infinite loop.

This patch declares the socket ID type as 'int' or 'unsigned int' by
referring to the common practice in EAL.

Fixes: 56b6ef874f80 ("efd: new Elastic Flow Distributor library")
Fixes: f00708c2aa53 ("node: add IPv4 rewrite and lookup control")
Cc: stable@dpdk.org

Signed-off-by: Chengchang Tang <tangchengchang@huawei.com>
Signed-off-by: Min Hu (Connor) <humin29@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>