net/hns3: select Tx prepare based on Tx offload
Tx prepare should be called only when necessary to reduce the impact on
performance.

For partial TX offload, users need to call rte_eth_tx_prepare() to invoke
the tx_prepare callback of PMDs. In this callback, the PMDs adjust the
packet based on the offloading used by the user. (e.g. For some PMDs,
pseudo-headers need to be calculated when the TX cksum is offloaded.)

However, for the users, they cannot grasp all the hardware and PMDs
characteristics. As a result, users cannot decide when they need to
actually call tx_prepare. Therefore, we should assume that the user calls
rte_eth_tx_prepare() when using any Tx offloading to ensure that related
functions work properly. Whether packets need to be adjusted should be
determined by PMDs. They can make judgments in the dev_configure or
queue_setup phase. When the related function is not used, the pointer of
tx_prepare should be set to NULL to reduce the performance loss caused by
invoking rte_eth_tx_repare().

In this patch, if tx_prepare is not required for the offloading used by
the users, the tx_prepare pointer will be set to NULL.

Fixes: bba636698316 ("net/hns3: support Rx/Tx and related operations")
Cc: stable@dpdk.org

Signed-off-by: Chengchang Tang <tangchengchang@huawei.com>
Signed-off-by: Min Hu (Connor) <humin29@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>